FILE: renames route.enum.ts to routes.ts in SI
