+++ libs/spirit-islander/shared/data-access/src/index.ts
@@ -1,3 +1,3 @@
-export * from './lib/route.enum';
+export * from './lib/routes';
 export * from './lib/router.service';
 export * from './lib/state.service';
